Remove an interface that causes an unnecessary cast and hides the true type of the object.
