--- framework/src/play/src/main/java/play/mvc/HandlerRef.java
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
- */
-package play.mvc;
-
-/**
- * Reference to an Handler.
- */
-public interface HandlerRef {
-    
-}
