+++ framework/src/play/src/main/scala/play/api/mvc/Action.scala
@@ -5,7 +5,6 @@ package play.api.mvc
 
 import play.api.libs.iteratee._
 import play.api._
-import play.core.Router.{ HandlerInvoker, HandlerInvokerFactory }
 import scala.concurrent._
 import scala.language.higherKinds
 
@@ -25,33 +24,6 @@ trait Handler
  */
 trait RequestTaggingHandler extends Handler {
   def tagRequest(request: RequestHeader): RequestHeader
-}
-
-/**
- * Reference to a Handler, useful for contructing handlers from Java code.
- */
-class HandlerRef[T](call: => T, handlerDef: play.core.Router.HandlerDef)(implicit hif: play.core.Router.HandlerInvokerFactory[T]) extends play.mvc.HandlerRef {
-
-  lazy val invoker: HandlerInvoker[T] = hif.createInvoker(call, handlerDef)
-
-  /**
-   * Retrieve a real handler behind this ref.
-   */
-  def handler: play.api.mvc.Handler = {
-    invoker.call(call)
-  }
-
-  /**
-   * String representation of this Handler.
-   */
-  lazy val sym = {
-    handlerDef.controller + "." + handlerDef.method + "(" + handlerDef.parameterTypes.map(_.getName).mkString(", ") + ")"
-  }
-
-  override def toString = {
-    "HandlerRef[" + sym + ")]"
-  }
-
 }
 
 /**
